backend/pyproject.toml
[tool.poetry]
name = "fido test"
version = "0.1.0"
description = "takhome"
authors = ["Daniel Eje <ejedenials@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
basicauth = "1.0.0"
fastapi = "0.95.0"
pydantic = "1.10.6"
starlette = "0.26.1"
uvicorn = "0.21.1"
gunicorn = "20.1.0"
passlib = "1.7.4"
fastapi-event = "0.1.3"
ujson = "5.7.0"
greenlet = "1.1.2"
requests-toolbelt = "0.10.1"
python-dotenv = "1.0.1"
pytest-asyncio = "^0.20.3"
pylint = "^2.16.2"
pytest = "^7.2.1"
faker = "^17.0.0"
pyyaml = "^6.0"
black = "^23.1.0"
click = "^8.1.3"
isort = "^5.12.0"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
